Create basic KomaMRIBase subpackage
