+++ KomaMRIBase/Project.toml
@@ -1,0 +1,21 @@
+name = "KomaMRIBase"
+uuid = "d0bc0b20-b151-4d03-b2a4-6ca51751cb9c"
+authors = ["Carlos Castillo Passi <cncastillo@uc.cl>"]
+version = "0.8.0"
+
+[deps]
+Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
+MRIBase = "f7771a9a-6e57-4e71-863b-6e4b6a2f17df"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+
+[compat]
+Interpolations = "0.13, 0.14"
+MAT = "0.10"
+MRIBase = "0.3, 0.4"
+Parameters = "0.12"
+Pkg = "1.4"
+Reexport = "1"
+julia = "1.6"
