+++ KomaMRIBase/src/datatypes/Scanner.jl
