+++ KomaMRIBase/src/datatypes/sequence/ADC.jl
