+++ KomaMRIBase/src/datatypes/simulation/DiffusionModel.jl
