+++ KomaMRIBase/src/timing/TrapezoidalIntegration.jl
