+++ KomaMRICore/src/simulation/SimulatorCore.jl
@@ -1,5 +1,4 @@
 abstract type SimulationMethod end #get all available types by using subtypes(KomaMRI.SimulationMethod)
-abstract type SpinStateRepresentation{T<:Real} end #get all available types by using subtypes(KomaMRI.SpinStateRepresentation)
 
 #Defined methods:
 include("Bloch/BlochSimulationMethod.jl") #Defines Bloch simulation method
