+++ KomaMRIFiles/test/generate_base_test_files/run.jl
@@ -6,7 +6,7 @@
 # * sphere_chemical_shift.jld2
 # * jemris_signals_epi_sphere_cs.jld2
 # * Koma_signal.jld2
-# in the directory KomaMRICore/test/test_files/
+# in the directory KomaMRIBase/test/test_files/
 #
 # Note that the saved phantom sphere_chemical_shift.jld2 doesn't have stored the information
 # of the motion functions
@@ -14,18 +14,18 @@
 using JLD2, HDF5, KomaMRIBase, KomaMRIFiles
 path = @__DIR__
 path_mrifiles = path*"/../test_files"
-path_mricore = path*"/../../../KomaMRICore/test/test_files"
-mkpath(path_mricore)
+path_mribase = path*"/../../../KomaMRIBase/test/test_files"
+mkpath(path_mribase)
 
 # Save sequence structs to files
 seq = read_seq(path_mrifiles*"/spiral.seq")
-save_object(path_mricore*"/spiral.jld2", seq)
+save_object(path_mribase*"/spiral.jld2", seq)
 seq = read_seq(path_mrifiles*"/epi.seq")
-save_object(path_mricore*"/epi.jld2", seq)
+save_object(path_mribase*"/epi.jld2", seq)
 seq = read_seq(path_mrifiles*"/radial_JEMRIS.seq")
-save_object(path_mricore*"/radial_JEMRIS.jld2", seq)
+save_object(path_mribase*"/radial_JEMRIS.jld2", seq)
 seq = read_seq(path_mrifiles*"/epi_100x100_TE100_FOV230.seq")
-save_object(path_mricore*"/epi_100x100_TE100_FOV230.jld2", seq)
+save_object(path_mribase*"/epi_100x100_TE100_FOV230.jld2", seq)
 
 # Save phantom structs to files
 # (there is a problem when reading the functions, so we don't save them)
@@ -36,14 +36,14 @@ function phantom_to_namedtuple(obj::Phantom)
 end
 obj = read_phantom_jemris(path_mrifiles*"/sphere_chemical_shift.h5")
 tuple_obj = phantom_to_namedtuple(obj)
-save_object(path_mricore*"/sphere_chemical_shift.jld2", tuple_obj)
+save_object(path_mribase*"/sphere_chemical_shift.jld2", tuple_obj)
 
 # Save jemris signal
 sig_jemris = h5open(path_mrifiles*"/jemris_signals_epi_sphere_cs.h5")["/signal/channels/00"]
 sig_jemris = sig_jemris[1,:] + 1im*sig_jemris[2,:]
 sig_jemris = sig_jemris[:]
-save_object(path_mricore*"/jemris_signals_epi_sphere_cs.jld2", sig_jemris)
+save_object(path_mribase*"/jemris_signals_epi_sphere_cs.jld2", sig_jemris)
 
 # Save Koma_signal
 raw = RawAcquisitionData(ISMRMRDFile(path_mrifiles*"/Koma_signal.mrd"))
-save_object(path_mricore*"/Koma_signal.jld2", raw)
+save_object(path_mribase*"/Koma_signal.jld2", raw)
