+++ src/KomaMRI.jl
@@ -8,6 +8,7 @@ using MRIReco
 
 # IMPORT PACKAGES
 using Reexport
+@reexport using KomaMRIBase
 @reexport using KomaMRICore
 @reexport using KomaMRIFiles
 @reexport using KomaMRIPlots
